+++ src/components/form/BulmaTextArea.vue
@@ -1,23 +1,28 @@
 <template>
-	<textarea class="textarea" :rows="rows" :class="classes" />
+	<textarea :class="classes" :value="modelValue" class="textarea"
+	          @input="emit('update:modelValue',$event.target.value)">
+	</textarea>
 </template>
 
 <script lang="ts" setup>
 	import {computed} from "vue";
-	import {BulmaColour, BulmaSize, getBulmaClassesFromProps} from "../../types";
+	import {BulmaColour, BulmaSize, BulmaState, getBulmaClassesFromProps} from "../../types";
 
-	//todo: add v:model support
-
+	const emit = defineEmits<{ (name: 'update:modelValue', value: string): void }>();
 	const props = defineProps<{
+		/** The value of the textarea */
+		modelValue?: string;
+		/** The colour of the textarea */
 		colour?: BulmaColour
+		/** The size of the textarea */
 		size?: BulmaSize
+		/** The visual state of the textarea */
 		//todo: loading needs to go on parent control tag
-		isLoading?: boolean
-		isStatic?: boolean
-		rows?: number
+		state?: BulmaState;
+		/** Static textarea size; can't be changed by user
+		 * @default false */
 		hasFixedSize?: boolean
 	}>();
 
 	const classes = computed(() => getBulmaClassesFromProps(props));
 </script>
-